Fix for bug #13363: ray (qmc) shadows had some light leaking issues, due to jittering of the start position for antialiasing in a pixel.
Now it distributes the start position over the fixed osa sample
positions, instead of of random positions in space. The ugly bit is
that a custom ordering was defined for osa 8/11/16 to ensure that the
first 4 are distributed relatively fair for adaptive sampling to decide
if more samples need to be taken.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@16960 954f8c5b-7b00-dc11-b283-0030488c597c