dpx/cineon was using the OUTPUT settings when loading DPX - white point, blackpoint, gamma, log. Allow DPX's with different color channel types to load, even if they are unsupported.
Saving log DPX's is broken, the header will always be written with log off because dpxCreate is called before the image presets are set. However there is no access to image presets before running dpxCreate() so log data is written but the header always sets it off.
Made a fix for this but couldn't test because it seems loading log images is also broken.

Enable verbose DPX output when '-d' debug flag is used.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@18896 954f8c5b-7b00-dc11-b283-0030488c597c