2.5+ file read provision for 2.49
I've managed to find a basic fail-safe method to prevent using the read
data with an OK popup. I then still will read the file, but wait for user
confirmation to continue, or just free the database.

The pupup tells the user that at least animation data will be lost.

For background renders of 2.5+ things are more tricky, so here I've added
a default failure to read, returning 0, which will typically gracefully
quit blender.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@19975 954f8c5b-7b00-dc11-b283-0030488c597c