Bugfix #18756
Texture nodes: on file load, the preview render signal was ignored.
Found out it ignores it all the way, inserting refreshes all over, not
using the 'afterqueue'. Will live with that for now, in 2.5 it's nicer
supported anyway.

Just added another refresh line on the proper signal to make it look nice
on file loads.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@20236 954f8c5b-7b00-dc11-b283-0030488c597c