== AVI ==
Only open AVI files with at least one supported video track type.
(This is the real fix for DV-Type1-AVIs, since the FCC of iads, that
is used there, wasn't detected. But the code will happily open AVI-files
with only audio tracks... So the real fix is: only try to open things, that
contains something we know how to handle :) )



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@21062 954f8c5b-7b00-dc11-b283-0030488c597c