Fixing scons compile on windows.
Since bli_threads.h now includes pthreads directly, we need to had instructions in SConscripts everywhere for proper include path.

Frankly, I feel like this should be done in a global manner and not in a per lib fashion, but that is for another day.

This commit also fixes more missing properties

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@23587 954f8c5b-7b00-dc11-b283-0030488c597c