Bugfix #19835: While playing animation, Render>Dimensions>FrameRate does not change animation speed
The timestep used for the playback timer now gets adjusted accordingly when the frames-per-second setting gets changed during playback (i.e. in the "animation_step" operator). This is not as ideal as only updating this when the framerate setting is changed, but using an appropriate update function for this failed miserably.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@24445 954f8c5b-7b00-dc11-b283-0030488c597c