BGE: replace mesh works for Soft Body (including reinstantiation of physics soft body mesh).
Even a static mesh can be used as replacement: the mesh
will be instantiated with the soft body settings of the
object. The position and orientation of the soft body
is preserved after the replacement.

Known limitation: the velocity of the soft body is reset
aftet the replacement. This is because soft body don't
have a well defined velocity.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@24871 954f8c5b-7b00-dc11-b283-0030488c597c