OSX: populate file dialog system & bookmark defaults paths with the finder's ones
Former implementation was to populate the "system" paths with the mounted paths. The issue is that OSX mounts some paths that are of no use such as /home (The home dirs reside in /Users).

When compiled with 10.5 SDK, it retrieves the folders you have in the finder and OSX standard open/save dialog.
When compiled with older SDKs, it assumes these folders are the default ones (home, Desktop, Documents, Pictures, Music, Movies).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@25430 954f8c5b-7b00-dc11-b283-0030488c597c