BGE bugfix: [#19476] Negative Lamps in BGE Do Not Function (in GLSL) this bug was introduced in Blender 2.49 (probably my own fault, in other part of the code though while fixing support for IPOs in GLSL Lamps).
The good news is: GLSL Lamps looks in BGE like working perfectly now.

Multitexture on the other hand is not supporting "negative" lamps (with the checkbox option on).

From my search (svn blame+log) it looks like multitexture never had negative lamp working. Actually in Blender 2.34 when this was introduced in BGE (maybe in Blender as well?) I couldn't see negative lamps working either. It's hard to test this with Blender 2.34 though (it even crash with my test file).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@25636 954f8c5b-7b00-dc11-b283-0030488c597c