Preview Range Tweak:
Made preview range be turned on/off using a proper flag instead of just relying on checking for start-frame = 0. It is no longer satisfactory to do that since we can have negative frame numbers, and also having it as a proper flag means that the range can be toggled on/off non-destructively.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26119 954f8c5b-7b00-dc11-b283-0030488c597c