Bugfix #19970: auto-clamped / auto working strangly in f-curve editor
Fixed the operators for DopeSheet/Graph Editors responsible for setting the "auto-clamped". This option is actually per F-Curve instead of per handle, and the code here should function like it did in 2.4x

However, despite this, it still appears to work oddly IMO. Any comments Bassam or animators familiar with the intentions of this?

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26510 954f8c5b-7b00-dc11-b283-0030488c597c