J-key render switching back, now with 10 slots.
Implementation note: this was done by giving each Render a slot number,
and for every slot a new Render will be created. Not sure if this is
ideal, but it ensures that all passes, render info, etc are separate so
you can also compare render layers and passes, in 2.4x only whatever it
was currently displaying was backed up.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26747 954f8c5b-7b00-dc11-b283-0030488c597c