Bugfix: 3DS Importer not settings UV's correctly
Thanks to example .3ds file and bug report from stridernzl.

1) The API function for adding textures to meshes was not working correctly, resulting in a Null texture slot being created (i.e. a texture slot was created, but the given texture could never be assigned to this). 
Fixed by making the function always require a texture as input (as with other places that use pointers as arguments).

2) In "Textured" draw mode, the imported meshes were shown white (i.e. without any textures). 
It appears that that the old old setting (mesh.faceUV) is no longer valid/wrapped in RNA, so worked around this by setting the 'tex' setting for UV-faces on. From the UI, this seems to do the same thing as entering editmode and assigning an image to the faces from the image editor.

---

I've also removed a few lines of commented 2.4x code that is no longer valid. Hopefully there wasn't anything too valuable that I accidentally removed in this process ;)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@26758 954f8c5b-7b00-dc11-b283-0030488c597c