Pose Tools Cleanup:
Moved some of the generic code used to determine the F-Curves linked to PoseChannel transforms (as used by the Pose Sliding tools) into a separate file, in preparation for migration of PoseLib tools to this system too. This should make it easier to add some useful new functionality to the PoseLib browsing system (pending in a later commit).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@27020 954f8c5b-7b00-dc11-b283-0030488c597c