Fix for [#21773] Cast Modifier cant use empties as centers
Based on the assumption that requiring object targets to be OB_EMPTY makes any other object compatible as a target. If the assumption is wrong can be reverted. Only the Cast modifier uses this at the moment and to me it looks like Cast only uses object transform so should be fine.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@27779 954f8c5b-7b00-dc11-b283-0030488c597c