Spline IK - Influence Control
Made the 'Influence' slider work for Spline IK too, and made that setting visible now that it works. 

Note that there is still some popping that can occur when going to/from influence = 0.0. I'm not sure exactly what's causing this yet, but hopefully it won't be too noticeable in practice.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@28368 954f8c5b-7b00-dc11-b283-0030488c597c